+++ lib/rubocop/config_store.rb
@@ -21,6 +21,7 @@ module Rubocop
 
     def set_options_config(options_config)
       loaded_config = ConfigLoader.load_file(options_config)
+      ConfigLoader.make_excludes_absolute(loaded_config)
       @options_config = ConfigLoader.merge_with_default(loaded_config,
                                                         options_config)
     end